Made code more efficient

+++ 007.cpp
@@ -3,8 +3,8 @@
 
 using namespace std;
 
-bool isPrime(long int number) {
-    for(long int j = 2; j < sqrt(number) + 1; j++) {
+bool isPrime(int number) {
+    for(int j = 2; j < sqrt(number) + 1; j++) {
         if(number % j == 0) {
             return false;
         }
@@ -16,7 +16,7 @@ int main() {
     int desired = 10001;
     int primes[desired];
     primes[0] = 2;
-    long int newNumber;
+    int newNumber;
     for(int i = 1; i <= desired; i++) {
         bool found = false;
         newNumber = primes[i-1] + 1;
